+++ src/net/sourceforge/kolmafia/CakeArenaManager.java
@@ -128,7 +128,7 @@ public class CakeArenaManager
 	public static final int earnedXP( final String response )
 	{
 		Matcher matcher = CakeArenaManager.WIN_PATTERN.matcher( response );
-		return matcher.find() ? Integer.valueOf( matcher.group( 1 ) ).intValue() : 0;
+		return matcher.find() ? Integer.valueOf( matcher.group( 1 ) ) : 0;
 	}
 
 	/**
