+++ src/net/sourceforge/kolmafia/request/QuestLogRequest.java
@@ -196,7 +196,7 @@ public class QuestLogRequest
 		{
 			Integer key = it.next();
 			String header = map.get( key );
-			String cut = responseText.substring( key.intValue() ).split( "</blockquote>" )[ 0 ];
+			String cut = responseText.substring( key ).split( "</blockquote>" )[ 0 ];
 
 			if ( header.equals( "Council Quests:" ) )
 			{
